// user-simulator/src/types/clients.ts
export interface Client {
	id: string;
	name: string;
	address: string;
	coords: Coordinates;
	is_active: boolean;
	created_at: Date;
	last_activity: Date;
	jobs: Job[];
	contacts: ClientContactLink[];
	notes: ClientNote[];
}
// ...
export interface ClientNote {
	id: string;
	client_id: string;
	content: string;
	created_at: Date;
	updated_at: Date;

	creator_tech_id: string | null;
	creator_dispatcher_id: string | null;
	creator_tech?: { id: string; name: string; email: string };
	creator_dispatcher?: { id: string; name: string; email: string };

	last_editor_tech_id: string | null;
	last_editor_dispatcher_id: string | null;
	last_editor_tech?: { id: string; name: string; email: string };
	last_editor_dispatcher?: { id: string; name: string; email: string };
}
// ...
export interface Contact {
	id: string;
	name: string;
	email: string | null;
	phone: string | null;
	type: string | null; // "customer", "vendor", "property_manager", etc.
	is_active: boolean;
	created_at: Date;
	updated_at: Date;

	client_contacts: ClientContactLink[];
}

export interface ClientContactLink {
	client_id: string;
	contact_id: string;
	relationship: string; // "owner", "tenant", "manager", "emergency_contact"
	is_primary: boolean;
	is_billing: boolean;
	created_at: Date;
	updated_at: Date;

	client?: Client;
	contact?: Contact;
}